Added parameter to only execute a single testcase

+++ configuration/buildAndUpdateProduct.sh
@@ -95,7 +95,7 @@ parallelexecution=0
 p2local=0
 
 if [ $# -eq 0 ]; then
-    echo "buildAndUpdateProduct [-b -u -g -t -a -r -o -c -p -v -m <config> -n <package> -l <port> -x <gwt-workers>] [build|install|all|hot-deploy|remote-deploy|local-deploy|release]"
+    echo "buildAndUpdateProduct [-b -u -g -t -a -r -o -c -p -v -m <config> -n <package> -l <port> -x <gwt-workers> -j <test-package>] [build|install|all|hot-deploy|remote-deploy|local-deploy|release]"
     echo ""
     echo "-g Disable GWT compile, no gwt files will be generated, old ones will be preserved."
     echo "-b Build GWT permutation only for one browser and English language."
@@ -115,6 +115,7 @@ if [ $# -eq 0 ]; then
     echo "-u Run without confirmation messages. Use with extreme care."
     echo "-v Build local p2 respository, and use this instead of p2.sapsailing.com"
     echo "-x <number-of-workers> use this many worker threads for building GWT permutations (default: 2)."
+    echo "-j <test-package> only execute the provided test package during tests"
     echo ""
     echo "build: builds the server code using Maven to $PROJECT_HOME (log to $START_DIR/build.log)"
     echo ""
@@ -151,7 +152,7 @@ echo SERVERS_HOME is $SERVERS_HOME
 echo BRANCH is $active_branch
 echo VERSION is $VERSION_INFO
 
-options=':bgtocparvm:n:l:s:w:x:u'
+options=':bgtocparvm:n:l:s:w:x:j:u'
 while getopts $options option
 do
     case $option in
@@ -172,6 +173,7 @@ do
         u) suppress_confirmation=1;;
         v) p2local=1;;
         x) GWT_WORKERS=$OPTARG;;
+        j) TESTCASE_TO_EXECUTE=$OPTARG;; 
         \?) echo "Invalid option"
             exit 4;;
     esac
@@ -535,7 +537,7 @@ if [[ "$@" == "build" ]] || [[ "$@" == "all" ]]; then
         fi
 
 	cd $PROJECT_HOME/java
-	if [ $gwtcompile -eq 1 ]; then
+	if [ $gwtcompile -eq 1 ] && [[ "$clean" == "clean" ]]; then
 	    echo "INFO: Compiling GWT (rm -rf com.sap.$PROJECT_TYPE.gwt.ui/com.sap.$PROJECT_TYPE.*)"
 	    rm -rf com.sap.$PROJECT_TYPE.gwt.ui/com.sap.$PROJECT_TYPE.*
         GWT_XML_FILES=`find com.sap.$PROJECT_TYPE.gwt.ui/src/main/resources -name '*.gwt.xml'`
@@ -587,6 +589,11 @@ if [[ "$@" == "build" ]] || [[ "$@" == "all" ]]; then
     else
         extra="$extra -DskipTests=false"
         # TODO: Think about http://maven.apache.org/surefire/maven-surefire-plugin/examples/fork-options-and-parallel-execution.html
+        if [[ "$TESTCASE_TO_EXECUTE" != "" ]]; then
+            # http://maven.apache.org/surefire/maven-surefire-plugin/examples/single-test.html
+            echo "Running only testcase $TESTCASE_TO_EXECUTE"
+            extra="$extra -Dtest=$TESTCASE_TO_EXECUTE"
+        fi
     fi
 
     if [ $android -eq 0 ] && [ $gwtcompile -eq 0 ] && [ $testing -eq 0 ]; then
